OpenCL formats: Use size_t instead of int in a bunch of places because of problems where eg. a GWS of 32M x buffer size of 64 ends up in 2G which doesn't fit in a signed int.

+++ src/common-opencl.c
@@ -59,7 +59,7 @@ static int number_of_events;
 static int * split_events;
 static cl_event * to_profile_event;
 static struct fmt_main * self;
-static void (*create_clobj)(int gws, struct fmt_main * self);
+static void (*create_clobj)(size_t gws, struct fmt_main * self);
 static void (*release_clobj)(void);
 static const char * config_name;
 static size_t gws_limit;
@@ -1027,7 +1027,7 @@ void opencl_init_auto_setup(
 	int p_default_value, int p_hash_loops, int p_number_of_events,
 	int * p_split_events, const char ** p_warnings,
 	cl_event * p_to_profile_event, struct fmt_main * p_self,
-	void (*p_create_clobj)(int gws, struct fmt_main * self),
+	void (*p_create_clobj)(size_t gws, struct fmt_main * self),
 	void (*p_release_clobj)(void), int p_buffer_size, size_t p_gws_limit)
 {
 	int i;
