+++ src/common-opencl.h
@@ -229,7 +229,7 @@ void opencl_init_auto_setup(
 	int p_default_value, int p_hash_loops, int p_number_of_events,
 	int * p_split_events, const char ** p_warnings,
 	cl_event * p_to_profile_event, struct fmt_main * p_self,
-	void (*p_create_clobj)(int gws, struct fmt_main * self),
+	void (*p_create_clobj)(size_t gws, struct fmt_main * self),
 	void (*p_release_clobj)(void), int p_buffer_size, size_t p_gws_limit);
 
 #endif
