+++ src/opencl_rawmd5_fmt.c
@@ -75,7 +75,7 @@ static struct fmt_tests tests[] = {
 	{NULL}
 };
 
-static void create_clobj(int kpc, struct fmt_main * self)
+static void create_clobj(size_t kpc, struct fmt_main * self)
 {
 	self->params.min_keys_per_crypt = self->params.max_keys_per_crypt = kpc;
