+++ src/opencl_rawsha256_fmt.c
@@ -108,7 +108,8 @@ static void crypt_one(int index, sha256_hash * hash) {
 }
 
 /* ------- Create and destroy necessary objects ------- */
-static void create_clobj(int gws, struct fmt_main * self) {
+static void create_clobj(size_t gws, struct fmt_main * self)
+{
 	self->params.min_keys_per_crypt = self->params.max_keys_per_crypt = gws;
 
 	pinned_saved_keys = clCreateBuffer(context[ocl_gpu_id],
